refactor: audit-driven correctness/hardening pass + RDNA4 (gfx1201) tuning
Correctness and hardening across autograd, ops, GPU kernels, and
untrusted-input parsing, plus RDNA4 wave32/WMMA performance tuning.
106/107 HIP tests pass (the one failure is the intentionally-guarded
fp16 path); an asserts-enabled rebuild shows 0 regressions vs main.

- Training loop/optimizer: removed sync-storm .item() skip heuristics;
  kept clip_grad_norm_ + a single NaN check
- Core/autograd: dtype-checked data_ptr<T>(), in-place autograd guards,
  allocator static-destruction fix, round-to-nearest-even fp16 conversion
- Ops/gradients: differentiable cat(), tie-splitting max/min backward,
  fused RMSNorm+Linear graph fix
- GPU kernels: online-softmax block-max broadcast, gather/scatter bounds
  guards, sampling top-k barrier + float atomicMax fixes, wave32
  reduction corrections
- Input hardening: safetensors/serialization/model-loader bounds and
  dtype-enum validation
- Features: ignore_index cross-entropy, beam-search KV-cache reorder,
  PrefetchDataLoader deadlock/exception fixes, deterministic RNG,
  ModuleList parameter traversal
- Dead-code removal + build hygiene (USE_CUDA gating, IWYU includes,
  .gitignore build/CTest artifacts)

+++ src/generation/generator.cpp
@@ -99,8 +99,8 @@ Tensor Generator::generate(const Tensor& prompt_ids, const SamplingParams& param
         // For GPU: transfer just the stop mask (small - batch_size bytes)
         Tensor stop_cpu = stop_mask.to(Device::CPU);
         Tensor done_cpu = done_mask.to(Device::CPU);
-        const int8_t* stop_data = stop_cpu.data_ptr<const int8_t>();
-        int8_t* done_data = done_cpu.data_ptr<int8_t>();
+        const int32_t* stop_data = stop_cpu.data_ptr<const int32_t>();
+        int32_t* done_data = done_cpu.data_ptr<int32_t>();
         for (int64_t b = 0; b < batch_size; ++b) {
             if (stop_data[b]) done_data[b] = 1;
         }
